requirements.txt
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
Django==2.1.4
djangorestframework==3.9.0
djangorestframework-simplejwt==3.3
kombu==4.2.2.post1
PyJWT==1.7.1
pytz==2018.7
vine==1.1.4
